+++ app/src/main/res/layout/fragment_create_unit_umum.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".CreateUnitUmumFragment"
+    tools:context=".unit.CreateUnitUmumFragment"
     android:paddingHorizontal="32dp"
     android:layout_marginTop="14dp"
     android:layout_marginBottom="35dp"
